bot/requirements.txt
nextcord
requests
pymongo
dnspython
validators
pygithub
python-dotenv
pillow
millify
nekos.py
aiohttp
algoliasearch